+++ frontend/src/App.jsx
@@ -11,7 +11,7 @@ import Home from "./pages/Home/Home.jsx";
 import LoginPage from "./pages/LoginSignup/LoginPage.jsx";
 import Signup from "./pages/LoginSignup/SignupPage.jsx";
 import VideoPreview from "./pages/VideoPreview/VideoPreview.jsx";
-
+import Contact from './pages/Contact/Contact.jsx'
 function App() {
   const clientId =
     "715277608713-sgtqudvfba9sr26avhok47041iskncu4.apps.googleusercontent.com";
@@ -25,7 +25,7 @@ function App() {
             <Route exact path="/login" element={<LoginPage />} />
             <Route exact path="/signup" element={<Signup />} />
             <Route exact path="/videopreview" element={<VideoPreview />} />
-
+            <Route exaxt path="/contact" element={ <Contact />} />
           </Routes>
         </div>
       </Router>
