+++ frontend/src/components/NavBar/NavBar.jsx
@@ -1,7 +1,6 @@
 import { Link } from "react-router-dom";
 import { Disclosure } from "@headlessui/react";
 import { useNavigate } from "react-router-dom";
-import { useState, useEffect } from "react";
 import ContactSupportIcon from "@mui/icons-material/ContactSupport";
 
 const NavBar = () => {
@@ -12,35 +11,6 @@ const NavBar = () => {
     navigate("/");
   };
 
-  const url = "http://localhost:5000/api/profile";
-  const [user, setUser] = useState({});
-
-  const fetchInfo = () => {
-    const authToken = localStorage.getItem("authToken"); // Get authToken from localStorage
-
-    return fetch(url, {
-      method: "GET",
-      headers: {
-        Authorization: `Bearer ${authToken}`,
-        "Content-Type": "application/json",
-      },
-    })
-      .then((res) => {
-        if (!res.ok) {
-          throw new Error("Network response was not ok");
-        }
-        return res.json();
-      })
-      .then((response) => setUser(response.user))
-      .catch((error) => {
-        console.error("Error fetching user data:", error);
-      });
-  };
-
-  useEffect(() => {
-    fetchInfo();
-  }, []);
-
   return (
     <div className={`w-full bg-dark text-light`}>
       <nav className="container relative flex flex-wrap items-center justify-between p-6 mx-auto lg:justify-between xl:px-0">
@@ -48,13 +18,6 @@ const NavBar = () => {
           {({ open }) => (
             <>
               <div className="flex flex-wrap items-center justify-between w-full lg:w-auto">
-                {localStorage.getItem("authToken") ? (
-                  <Link className="navbar-brand" to="/profile">
-                    <ul>{user.username}</ul>
-                  </Link>
-                ) : (
-                  ""
-                )}
                 <Link to="/">
                   <span className="flex items-center space-x-2 text-2xl font-medium text-indigo-500 dark:text-gray-100 font-sans ml-2">
                     <span>FinancialFriend</span>
@@ -63,7 +26,7 @@ const NavBar = () => {
 
                 <Disclosure.Button
                   aria-label="Toggle Menu"
-                  className="px-2 py-1 ml-auto text-gray-500 rounded-md lg:hidden hover:text-indigo-500 focus:text-indigo-500 focus:bg-indigo-100 focus:outline-none dark:text-gray-300 dark:focus:bg-trueGray-700"
+                  className="px-2 py-1 ml-auto text-neutral-400 rounded-md lg:hidden hover:text-indigo-500 focus:text-indigo-500 focus:bg-indigo-100 focus:outline-none dark:text-gray-300 dark:focus:bg-trueGray-700"
                 >
                   <svg
                     className="w-6 h-6 fill-current"
@@ -91,32 +54,32 @@ const NavBar = () => {
                     <div>
                       <Link
                         to="/"
-                        className="w-full px-4 py-2 -ml-4 text-gray-500 rounded-md dark:text-gray-300 hover:text-indigo-500 focus:text-indigo-500 focus:bg-indigo-100 dark:focus:bg-gray-800 focus:outline-none"
+                        className="w-full px-4 py-2 -ml-4 text-neutral-400 rounded-md dark:text-gray-300 hover:text-indigo-500 focus:text-indigo-500 focus:bg-indigo-100 dark:focus:bg-gray-800 focus:outline-none"
                       >
                         Home
                       </Link>
                       <Link
                         to="/article"
-                        className="w-full px-4 py-2 -ml-4 text-gray-500 rounded-md dark:text-gray-300 hover:text-indigo-500 focus:text-indigo-500 focus:bg-indigo-100 dark:focus:bg-gray-800 focus:outline-none"
+                        className="w-full px-4 py-2 -ml-4 text-neutral-400 rounded-md dark:text-gray-300 hover:text-indigo-500 focus:text-indigo-500 focus:bg-indigo-100 dark:focus:bg-gray-800 focus:outline-none"
                       >
                         Article
                       </Link>
                       <Link
                         to="/community"
-                        className="w-full px-4 py-2 -ml-4 text-gray-500 rounded-md dark:text-gray-300 hover:text-indigo-500 focus:text-indigo-500 focus:bg-indigo-100 dark:focus:bg-gray-800 focus:outline-none"
+                        className="w-full px-4 py-2 -ml-4 text-neutral-400 rounded-md dark:text-gray-300 hover:text-indigo-500 focus:text-indigo-500 focus:bg-indigo-100 dark:focus:bg-gray-800 focus:outline-none"
                       >
                         Chat
-                      </Link>
-                      <Link
-                        to="/"
-                        className="w-full px-4 py-2 -ml-4 text-gray-500 rounded-md dark:text-gray-300 hover:text-indigo-500 focus:text-indigo-500 focus:bg-indigo-100 dark:focus:bg-gray-800 focus:outline-none"
-                      >
-                        <ContactSupportIcon fontSize="medium" />
                       </Link>
                     </div>
 
                     {!localStorage.getItem("authToken") ? (
                       <div className="d-flex">
+                        <Link
+                          to="/contact"
+                          className="w-full px-4 py-2 -ml-4 text-neutral-400 rounded-md dark:text-gray-300 hover:text-indigo-500 focus:text-indigo-500 focus:bg-indigo-100 dark:focus:bg-gray-800 focus:outline-none"
+                        >
+                          <ContactSupportIcon fontSize="medium" />
+                        </Link>
                         <Link
                           to="/login"
                           className="w-full px-4 mt-auto text-center text-white bg-indigo-600 rounded-md lg:ml-5"
@@ -126,6 +89,12 @@ const NavBar = () => {
                       </div>
                     ) : (
                       <div className="d-flex">
+                        <Link
+                          to="/contact"
+                          className="w-full px-4 py-2 -ml-4 text-neutral-400 rounded-md dark:text-gray-300 hover:text-indigo-500 focus:text-indigo-500 focus:bg-indigo-100 dark:focus:bg-gray-800 focus:outline-none"
+                        >
+                          <ContactSupportIcon fontSize="medium" />
+                        </Link>
                         <div
                           className="w-full px-4 mt-auto text-center text-white bg-indigo-600 rounded-md lg:ml-5"
                           onClick={handleLogout}
@@ -146,27 +115,21 @@ const NavBar = () => {
             <li className="mr-3 nav__item">
               <Link
                 to="/"
-                className="w-full px-4 py-2 -ml-4 text-gray-500 rounded-md dark:text-gray-300 hover:text-indigo-500 focus:text-indigo-500 focus:bg-indigo-100 dark:focus:bg-gray-800 focus:outline-none"
+                className="w-full px-4 py-2 -ml-4 text-neutral-400 rounded-md dark:text-gray-300 hover:text-indigo-500 focus:text-indigo-500 focus:bg-indigo-100 dark:focus:bg-gray-800 focus:outline-none"
               >
                 Home
               </Link>
               <Link
                 to="/article"
-                className="w-full px-4 py-2 -ml-4 text-gray-500 rounded-md dark:text-gray-300 hover:text-indigo-500 focus:text-indigo-500 focus:bg-indigo-100 dark:focus:bg-gray-800 focus:outline-none"
+                className="w-full px-4 py-2 -ml-4 text-neutral-400 rounded-md dark:text-gray-300 hover:text-indigo-500 focus:text-indigo-500 focus:bg-indigo-100 dark:focus:bg-gray-800 focus:outline-none"
               >
                 Article
               </Link>
               <Link
                 to="/community"
-                className="w-full px-4 py-2 -ml-4 text-gray-500 rounded-md dark:text-gray-300 hover:text-indigo-500 focus:text-indigo-500 focus:bg-indigo-100 dark:focus:bg-gray-800 focus:outline-none"
+                className="w-full px-4 py-2 -ml-4 text-neutral-400 rounded-md dark:text-gray-300 hover:text-indigo-500 focus:text-indigo-500 focus:bg-indigo-100 dark:focus:bg-gray-800 focus:outline-none"
               >
                 Chat
-              </Link>
-              <Link
-                to="/"
-                className="w-full px-4 py-2 -ml-4 text-gray-500 rounded-md dark:text-gray-300 hover:text-indigo-500 focus:text-indigo-500 focus:bg-indigo-100 dark:focus:bg-gray-800 focus:outline-none"
-              >
-                <ContactSupportIcon fontSize="medium" />
               </Link>
             </li>
           </ul>
@@ -176,6 +139,12 @@ const NavBar = () => {
           {!localStorage.getItem("authToken") ? (
             <div className="d-flex">
               <Link
+                to="/contact"
+                className="w-full px-4 py-2 -ml-4 text-neutral-400 rounded-md dark:text-gray-300 hover:text-indigo-500 focus:text-indigo-500 focus:bg-indigo-100 dark:focus:bg-gray-800 focus:outline-none"
+              >
+                <ContactSupportIcon fontSize="medium" />
+              </Link>
+              <Link
                 to="/login"
                 className="w-full px-4 py-2 mt-auto text-center text-white bg-indigo-600 rounded-md lg:ml-5"
               >
@@ -184,6 +153,12 @@ const NavBar = () => {
             </div>
           ) : (
             <div className="d-flex">
+              <Link
+                to="/contact"
+                className="w-full px-4 py-2 -ml-4 text-neutral-400 rounded-md dark:text-gray-300 hover:text-indigo-500 focus:text-indigo-500 focus:bg-indigo-100 dark:focus:bg-gray-800 focus:outline-none"
+              >
+                <ContactSupportIcon fontSize="medium" />
+              </Link>
               <div
                 className="w-full px-4 py-2 mt-auto text-center text-white bg-indigo-600 rounded-md lg:ml-5"
                 onClick={handleLogout}
